requirements.txt
arrow==1.3.0
asgiref==3.8.1
backports.functools-lru-cache==2.0.0
binaryornot==0.4.4
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
crispy-tailwind==1.0.3
Django==5.1
django-appconf==1.0.6
django-browser-reload==1.15.0
django-compressor==4.5.1
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-tailwind==3.8.0
djangorestframework==3.15.2
dotenv==0.9.9
faiss-cpu==1.11.0.post1
Faker==33.1.0
ffmpeg==1.4
fuzzywuzzy==0.18.0
idna==3.8
Jinja2==3.1.4
joblib==1.4.2
Levenshtein==0.27.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient==2.2.7
numpy==2.2.0
packaging==25.0
pandas==2.2.3
pillow==10.4.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.12.1
rcssmin==1.1.2
regex==2024.11.6
requests==2.32.3
rich==13.8.0
rjsmin==1.2.2
roman==5.0
scikit-learn==1.6.0
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
threadpoolctl==3.5.0
toml==0.10.2
typer==0.15.2
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
